skills/ppt-deck-builder/scripts/requirements.txt
requests
python-pptx
Pillow
